requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
eventlet==0.35.1
Werkzeug==3.0.1
undetected-chromedriver==3.5.5
selenium==4.15.2
SpeechRecognition==3.10.1
PyAudio==0.2.14
pydub==0.25.1
numpy==1.26.2
